# front/openapi/openApi.public.yaml
openapi: 3.0.1
info:
  title: Bank API — public (no auth routes)
  description: >
    Все пути кроме `/user-service/auth/*`. Для user/admin и BFF без операций входа/регистрации.
    Полный контракт — openApi.backend-gateway.yaml; только auth — openApi.sso.yaml.
  version: "1.0"
servers:
  - url: http://localhost:8000/api
    description: >
      Base URL BFF (не шлюз бекенда). В проде подставьте origin своего BFF.
security:
  - BffSessionCookie: []
paths:
...
  /core-api/cardaccount/open/{userId}:
    post:
      tags:
        - card-account-controller
      operationId: openAccount
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardAccountCreateModelDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /core-api/cardaccount/close/{accountId}:
    post:
      tags:
        - card-account-controller
      operationId: closeAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /core-api/cardaccount/{accountId}/set-main:
    post:
      tags:
        - card-account-controller
      operationId: setMainAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /core-api/cardaccount/{accountId}:
    get:
      tags:
        - card-account-controller
      operationId: getUserCardAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /core-api/cardaccount/exists/{accountId}:
    get:
      tags:
        - card-account-controller
      operationId: checkAccountExists
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /core-api/cardaccount/all/{userId}:
    get:
      tags:
        - card-account-controller
      operationId: getUserCardAccounts
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageIndex
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # Credit Service Endpoints
...
  /credit-service/credit/{userId}/get_by_user_id:
    get:
      tags:
        - credit-controller
      operationId: getByUserId
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credit'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /credit-service/credit_rating/{userId}/get_by_user:
    get:
      tags:
        - credit-rating-controller
      operationId: getCreditRatingByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreditRatingDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Currency:
      type: string
      enum:
        - USD
        - EUR
        - RUB
      default: RUB
...
    MoneyValueDto:
      type: object
      properties:
        value:
          type: number
        currency:
          $ref: '#/components/schemas/Currency'
...
    CardAccount:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          description: Отображаемое имя счёта
        main:
          type: boolean
          description: Главный счёт пользователя
        visible:
          type: boolean
          description: Видимость в списках (false = hidden)
        money:
          $ref: '#/components/schemas/MoneyValueDto'
        deleted:
          type: boolean
        transactionOperations:
          type: array
          items:
            $ref: '#/components/schemas/TransactionOperation'
    CardAccountCreateModelDto:
      type: object
      properties:
        name:
          type: string
        currency:
          $ref: '#/components/schemas/Currency'
    TransactionOperation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cardAccountId:
          type: string
          format: uuid
        dateTime:
          type: string
          format: date-time
        transactionType:
          type: string
          enum:
            - WITHDRAWAL
            - ENROLLMENT
        transactionActoin:
          type: string
        transactionStatus:
          type: string
          enum:
            - COMPLETE
            - IN_PROGRESS
            - DECLINED
        money:
          $ref: '#/components/schemas/MoneyValueDto'
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageCardAccount:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardAccount'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    # Credit Service Schemas
...
    CreditRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        percentageStrategy:
          type: string
          enum:
            - FROM_REMAINING_DEBT
            - FROM_TOTAL_DEBT
        collectionPeriodSeconds:
          type: integer
          format: int32
        openingDate:
          type: string
          format: date-time
          description: >
            Дата-время без указания зоны (как Java LocalDateTime), например 2007-12-03T10:15:30.
        ruleName:
          type: string
        percentage:
          type: number
    CreditRatingDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        rating:
          type: number
    Credit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        cardAccount:
          type: string
          format: uuid
        lastInterestUpdate:
          type: string
          format: date-time
        currentDebtSum:
          type: number
        initialDebt:
          type: number
        interestDebtSum:
          type: number
        currency:
          $ref: '#/components/schemas/Currency'
        creditRule:
          $ref: '#/components/schemas/CreditRule'
...
    BffErrorBody:
      type: object
      description: Обобщённое тело ошибки (BFF/шлюз; фактический JSON может отличаться).
      properties:
        message:
          type: string
        code:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    BadRequest:
      description: Некорректный запрос (синтаксис, параметры пути/запроса).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    Unauthorized:
      description: Нет или недействительная аутентификация (сессия, JWT, API-ключ).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    Forbidden:
      description: Доступ запрещён (недостаточно прав).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    UnprocessableEntity:
      description: Ошибка валидации тела запроса или бизнес-правил.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
    InternalServerError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BffErrorBody'
  securitySchemes:
    BffSessionCookie:
      type: apiKey
      in: cookie
      name: fins_session
      description: >
        Сессия браузера с BFF: httpOnly (и при необходимости Secure / SameSite) — имя куки задаёт BFF.
        Плейсхолдер `fins_session` замените на фактическое имя в своём BFF. Браузер не получает JWT бекенда.
    BffServiceApiKey:
      type: apiKey
      in: header
      name: X-API-KEY
      description: Сервисный ключ BFF при вызове internal-маршрутов (не сессия пользователя).